# pkg/mgmt/openapi.json
{
  "openapi": "3.0.3",
  "info": {"title": "OpenAPI Mock Management API", "version": "1.0.0"},
  "paths": {
    "/logs": {
      "get": {"responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/CallRecord"}}}}}}},
      "delete": {"responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusResponse"}}}}}}
    },
    "/logs/{request_id}": {
      "get": {
        "parameters": [{"name": "request_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/CallRecord"}}}}}}
      }
    },
    "/context-values": {
      "get": {"responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContextValuesStore"}}}}}},
      "put": {"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContextValuesStore"}}}}, "responses": {"200": {"description": "OK"}}},
      "patch": {"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContextValuesStore"}}}}, "responses": {"200": {"description": "OK"}}},
      "delete": {"responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusResponse"}}}}}}
    },
    "/context-values/{request_id}": {
      "parameters": [{"name": "request_id", "in": "path", "required": true, "schema": {"type": "string"}}],
      "get": {"responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContextValueObject"}}}}}},
      "put": {"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContextValueObject"}}}}, "responses": {"200": {"description": "OK"}}},
      "patch": {"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContextValueObject"}}}}, "responses": {"200": {"description": "OK"}}},
      "delete": {
        "requestBody": {"required": false, "content": {"application/json": {"schema": {"type": "object", "properties": {"keys": {"type": "array", "items": {"type": "string"}}}}}}},
        "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusResponse"}}}}}
      }
    },
    "/doc": {"get": {"responses": {"200": {"description": "OK"}}}},
    "/docs": {
      "get": {"responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MockDocsList"}}}}}}
    },
    "/docs/{api_name}": {
      "get": {
        "parameters": [{"name": "api_name", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "OK", "content": {"text/html": {"schema": {"type": "string"}}, "application/json": {"schema": {"$ref": "#/components/schemas/MockDocsVersions"}}}}, "404": {"description": "Not found"}}
      }
    },
    "/docs/{api_name}/openapi.json": {
      "get": {
        "parameters": [{"name": "api_name", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"type": "object"}}}}, "404": {"description": "Not found"}}
      }
    },
    "/docs/{api_name}/{api_ver}": {
      "get": {
        "parameters": [{"name": "api_name", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "api_ver", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "OK", "content": {"text/html": {"schema": {"type": "string"}}}}, "404": {"description": "Not found"}}
      }
    },
    "/docs/{api_name}/{api_ver}/openapi.json": {
      "get": {
        "parameters": [{"name": "api_name", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "api_ver", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"type": "object"}}}}, "404": {"description": "Not found"}}
      }
    },
    "/openapi.json": {"get": {"responses": {"200": {"description": "OK"}}}}
  },
  "components": {
    "schemas": {
      "CallRecord": {
        "type": "object",
        "properties": {
          "request_id": {"type": "string"},
          "method": {"type": "string"},
          "timestamp": {"type": "string", "format": "date-time"},
          "request": {},
          "response": {},
          "error": {"type": "string"},
          "panic": {"type": "string"},
          "duration_ms": {"type": "integer", "format": "int64"}
        }
      },
      "ContextValueObject": {"type": "object", "additionalProperties": true},
      "ContextValuesStore": {"type": "object", "additionalProperties": {"$ref": "#/components/schemas/ContextValueObject"}},
      "MockDocInfo": {
        "type": "object",
        "properties": {
          "api_name": {"type": "string"},
          "api_ver": {"type": "string"},
          "title": {"type": "string"},
          "ui_url": {"type": "string"},
          "openapi_url": {"type": "string"}
        }
      },
      "MockDocsList": {"type": "object", "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/MockDocInfo"}}}},
      "MockDocsVersions": {"type": "object", "properties": {"api_name": {"type": "string"}, "versions": {"type": "array", "items": {"$ref": "#/components/schemas/MockDocInfo"}}}},
      "StatusResponse": {"type": "object", "properties": {"status": {"type": "string"}}}
    }
  }
}